requirements.txt
# For UI
streamlit

# Langchain library
langchain

# For LLM
openai
langchain_openai

# Extras
python-dotenv
pydantic[email]

# For Tests
pytest